Rename Newsletters\ViewInBrowser to ViewInBrowserRenderer
[MAILPOET-2430]
